pyproject.toml
[tool.poetry]
name = "autotrader"
version = "0.1.0"
description = "Automated Trading Platform"
authors = ["Thomas Campion <campion.t@husky.neu.edu>"]

[tool.poetry.dependencies]
python = "3.7.4"
websocket-client = "^0.57.0"
tdameritrade = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
